+++ school_class.rb
@@ -1,0 +1,43 @@
+# This will be the SUPER CLASS for the STUDENT, TEACHER, and APPRENTICE TEACHER class
+
+# STUDENT, TEACHER, and APPRENTICE TEACHER all have the following ATTRIBUTES:
+  # :age
+  # :phase
+  # :name
+
+# STUDENT, TEACHER and APPRENTICE TEACHER all have the following shared METHODS:
+
+  # offer_high_five
+  # set_phase
+
+# Don't forget about CONSTANTS yo... => TEACHER_MAX_HEATH = 100 => self.class::TEACHER_MAX_HEALTH
+
+class School
+  attr_reader :age, :phase
+  attr_accessor :name
+
+  def initialize
+        @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+  end
+
+   def offer_high_five
+    "High five!"
+  end
+
+  def set_phase(num)
+    response = ""
+    if num == @phase
+      response += "I'm doing phase #{@phase} again because "
+      response += "I put my learning first. I'm gonna rock it!"
+    else
+      response = "Oooh, phase #{num}. I hope I'm ready!"
+    end
+    @phase = num
+    response
+  end
+
+
+
+
+end
