+++ student.rb
@@ -1,4 +1,6 @@
-class Student
+require_relative 'school_class'
+
+class Student < School
   attr_reader :age, :phase
   attr_accessor :name
 
@@ -8,21 +10,7 @@ class Student
     @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
 
-  def set_phase(num)
-    response = ""
-    if num == @phase
-      response += "I'm doing phase #{@phase} again because "
-      response += "I put my learning first. I'm gonna rock it!"
-    else
-      response = "Oooh, phase #{num}. I hope I'm ready!"
-    end
-    @phase = num
-    response
-  end
 
   def learn_stuff
     response = ""
